#40-UserRepo определил пакет для перечислений, переименовал файл со скриптом, изменил тип для полей связанных с id, добавил путь для хранения image и определил поведение liquibase в файле application.properties

+++ src/main/java/ru/skypro/homework/dto/user/User.java
@@ -2,11 +2,12 @@ package ru.skypro.homework.dto.user;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
+import ru.skypro.homework.enums.Role;
 
 @Data
 @Accessors(chain = true)
 public class User {
-    private long id;
+    private int id;
     private String email;
     private String firstName;
     private String lastName;
